Static & Dynamic Web View
In android, WebView is an extension of View class and it is used to show the static HTML web pages content or remote web pages content with URL in android applications as a part of our activity layout.

Generally, in android, WebView will act as an embedded browser to include the web pages content in our activity layout and it won’t contain any features of normal browsers, such as address bar, navigation controls, etc.