Grid Layout with Example
In android, Grid View is a ViewGroup that is used to display items in a two dimensional, scrollable grid and grid items are automatically inserted to the gridview layout using a list adapter.

Generally, the adapter pulls data from sources such as an array or database and converts each item into a result view and that’s placed into the list.