Android LifeCycle & Removing Status Bar
Activity Lifecycle: Activity is one of the building blocks of Android OS. In simple words
Activity is a screen that user interact with. Every Activity in android has lifecycle like created,
started, resumed, paused, stopped or destroyed. These different states are known as Activity
Lifecycle. In other words we can say Activity is a class pre-written in Java Programming.
Below is Activity Lifecycle Table:
Short description of Activity Lifecycle example:
onCreate() – Called when the activity is first created
onStart() – Called just after it’s creation or by restart method after onStop(). Here Activity start
becoming visible to user
onResume() – Called when Activity is visible to user and user can interact with it
onPause() – Called when Activity content is not visible because user resume previous activity
onStop() – Called when activity is not visible to user because some other activity takes place of
it
onRestart() – Called when user comes on screen or resume the activity which was stopped
onDestroy() – Called when Activity is not in background