RatingBar - added
In android, RatingBar is a UI control that is used to get the rating from the user. The RatingBar is an extension of SeekBar and ProgressBar that shows a rating in stars and it allows users to set the rating value by touch or click on the stars.

The android RatingBar will always return a rating value as a floating-point number such as 1.0, 2.0, 2.5, 3.0, 3.5, etc.