SeekBar - added
By using the draggable thumb in SeekBar we can slide left or right to choose a value between 0 and maximum integer value which we defined using android:max attribute. An example of SeekBar is our device Brightness control or volume control.

In android, by using SeekBar.OnSeekBarChangeListener listener, we can notify the client when the progress level of seekbar has been changed.