Basic Audio Player
In android, by using MediaPlayer class we can easily fetch, decode and play both audio and video files with minimal setup.

The android media framework provides built-in support for playing a variety of common media types, such as audio or video. We have multiple ways to play audio or video but the most important component of media framework is MediaPlayer class.

MediaPlayer class provides a different type of methods to control audio and video files based on requirements.

Method	Description
getCurrentPosition()	It is used to get the current position of the song in milliseconds.
getDuration()	It is used to get the total time duration of the song in milliseconds.
isPlaying()	It returns true / false to indicate whether song playing or not.
pause()	It is used to pause the song playing.
setAudioStreamType()	it is used to specify the audio streaming type.
setDataSource()	It is used to specify the path of audio / video file to play.
setVolume()	It is used to adjust media player volume either up / down.
seekTo(position)	It is used to move song to particular position in milliseconds.
getTrackInfo()	It returns an array of track information.
start()	It is used to start playing the audio/video.
stop()	It is used to stop playing the audio/video.
reset()	It is used to reset the MediaPlayer object.
release()	It is used to releases the resources which are associated with MediaPlayer object.