VideoView
In android, by using VideoView component and MediaController class we can easily implement the video player in android applications to play the videos with multiple playback options, such as play, pause, forward, backward, etc.

Generally, the MediaController class in android will provide playback options for video player, such as play, pause, backward, forward, etc.

The VideoView class in android will provide the functionality to fetch and play the videos using video player with minimal setup in android applications.

 the VideoView class provides a different type of methods to control video files based on requirements.

Method	Description
setMediaController()	It is used to set the media controller to videoview.
setVideoURI()	It is used to set the path of video file.
pause()	It is used to pause the video playing.
stopPlayback()	it is used to stop the video playing.
seekTo(position)	It is used to move video to a particular position in milliseconds.
resume()	It is used to resume the video playing.
start()	It is used to start playing the audio/video.
stopPlayback()	It is used to stop playing the audio/video.
Now we will see how to implement video playing applications using MediaController and VideoView with multiple playback options, such as play, pause, forward, backward in android application