Audio Recorder - IMPORTANT
In android, MediaRecorder class will provide a functionality to record audio or video files.

The android multimedia framework provides built-in support for capturing and encoding a variety of common audio and video formats. We have a multiple ways to record audio or video but by using MediaRecorder class we can easily implement audio or video recording.

In android, to record an audio we need to use device’s microphone along with MediaRecorder class. In case, if we want to record video, we need to use device’s camera along with MediaRecorder class.

MediaRecorder class provides a different type of methods to control audio and video recording based on requirements.

Method	Description
setAudioSource()	It is used to specify the source of audio to be recorded.
setVideoSource()	It is used to specify the source of the video to be recorded.
setOutputFormat()	It is used to specify the audio / video output format.
setAudioEncoder()	It is used to specify the audio encoder.
setVideoEncoder()	it is used to specify the video encoder.
setOutputFile()	It is used to specify the path of a recorded audio/video files to be stored.
stop()	It is used to stop the recording process.
start()	it is used to start the recording process.
release()	It is used to releases the resources which are associated with MediaRecorder object.