Services in android
In android, Service is a component which keep an app running in the background to perform long-running operations based on our requirements. For Service, we don’t have any user interface and it will run the apps in the background like playing the music in the background or handle network operations when the user in a different app.

Android Service Life Cycle
In android, the life cycle of service will follow two different paths Started or Bound.

Started Service
A service is Started when an application component, such as an activity calls startService() method. Once it started, it will run indefinitely in background even if the component that started is destroyed.

We can stop the Started service by using stopService() method or the service can stop itself by calling stopSelf() method. In android, the Started service component will perform a single operation and it won’t return any result to the caller.

Bound Service
A service is Bound when another application component calls bindService() method. The bound service runs as long as another application component is bound to it.

We can unbind the service by calling unbindService() method based on our requirements. In android, we can bind multiple components to a single service at once, but the service will be destroyed in case all the components unbind.